splitted search header with header
