+++ hw/code/ui/locators/base_locators.py
@@ -1,0 +1,5 @@
+from selenium.webdriver.common.by import By
+
+
+class BaseLocators():
+    HEADER_ERROR_MESSAGE = (By.ID, 'header_error-message')
